complete: data ingestion

+++ source/component/data_ingestion.py
@@ -19,21 +19,24 @@ class DataIngestion:
             if key == 'train':
                 collection_name = self.utility_config.train_collection_name
                 feature_store_file_path = self.utility_config.train_feature_store_dir_path
+                feature_store_file_name = self.utility_config.train_feature_store_file_name
             else:
                 collection_name = self.utility_config.predict_collection_name
                 feature_store_file_path = self.utility_config.predict_di_feature_store_file_path
+                feature_store_file_name = self.utility_config.predict_di_feature_store_file_name
 
             client = MongoClient(self.utility_config.mongodb_url_key)
             database = client[self.utility_config.database_name]
 
             collection = database[collection_name]
             cursor = collection.find()
-
             data = pd.DataFrame(list(cursor))
 
-            dir_path = os.path.dirname(feature_store_file_path)
-            os.makedirs(dir_path, exist_ok=True)
-            data.to_csv(feature_store_file_path, index=False)
+            # dir_path = os.path.dirname(feature_store_file_path)
+            # os.makedirs(dir_path, exist_ok=True)
+            # data.to_csv(feature_store_file_path, index=False)
+
+            export_data_csv(data, feature_store_file_name, feature_store_file_path)
 
             logging.info("complete: data load from mongoDB")
 
@@ -43,41 +46,40 @@ class DataIngestion:
             logging.error(e)
             raise e
 
-    def split_data_test_train(self, data: DataFrame) -> None:
+    def split_data_test_train(self, data: DataFrame):
         try:
             logging.info("start: train, test data split")
 
             train_set, test_set = train_test_split(data, test_size=self.utility_config.train_test_split_ratio, random_state=42)
 
-            dir_path = os.path.dirname(self.utility_config.train_file_path)
-            os.makedirs(dir_path, exist_ok=True)
-
-            train_set.to_csv(self.utility_config.train_file_path, index=False)
-            test_set.to_csv(self.utility_config.test_file_path, index=False)
-
             logging.info("complete: train, test data split")
+
+            return train_set, test_set
+
         except ChurnException as e:
             raise e
 
-    def clean_data(self, data):
+    def clean_data(self, data, key):
         try:
 
             logging.info("start: clean data")
 
-            data = data.drop_duplicates()
+            if key == 'train':
+                data = data.drop_duplicates()
 
-            data = data.loc[:, data.nunique() > 1]
+                data = data.loc[:, data.nunique() > 1]
 
-            drop_column = []
+                drop_column = []
 
-            for col in data.select_dtypes(include=['object']).columns:
-                unique_count = data[col].nunique()
+                for col in data.select_dtypes(include=['object']).columns:
+                    unique_count = data[col].nunique()
 
-                if unique_count / len(data) > 0.5:
-                    data.drop(col, axis=1, inplace=True)
-                    drop_column.append(col)
+                    if unique_count / len(data) > 0.5:
+                        data.drop(col, axis=1, inplace=True)
+                        drop_column.append(col)
 
-            logging.info(f"dropped columns: {drop_column}")
+                logging.info(f"dropped columns: {drop_column}")
+
             logging.info("complete: clean data")
 
             return data
@@ -91,7 +93,8 @@ class DataIngestion:
 
             if key == 'train':
                 mandatory_cols = self.utility_config.mandatory_col_list.copy()
-            else:
+
+            if key == 'predict':
                 mandatory_cols = self.utility_config.mandatory_col_list.copy()
                 mandatory_cols.remove(self.utility_config.target_column)
                 data = data.drop(self.utility_config.di_col_drop_in_clean, axis=1)
@@ -107,6 +110,8 @@ class DataIngestion:
                     except ValueError as e:
                         raise ChurnException(f"ERROR: converting data type for column: {col}")
 
+            data = data[mandatory_cols]
+
             logging.info("complete: process data")
 
             return data
@@ -115,16 +120,22 @@ class DataIngestion:
             raise e
 
     def initiate_data_ingestion(self, key):
+        try:
 
-        data = self.export_data_into_feature_store(key)
+            logging.info("start: data ingestion")
 
-        if key == 'train':
-            data = self.clean_data(data)
+            data = self.export_data_into_feature_store(key)
+            data = self.process_data(data, key)
+            data = self.clean_data(data, key)
 
-        data = self.process_data(data, key)
+            if key == 'train':
+                train_data, test_data = self.split_data_test_train(data)
+                export_data_csv(train_data, self.utility_config.train_file_name, self.utility_config.train_di_train_file_path)
+                export_data_csv(test_data, self.utility_config.test_file_name, self.utility_config.train_di_test_file_path)
 
-        if key == 'train':
-            self.split_data_test_train(data)
+            if key == 'predict':
+                export_data_csv(data, self.utility_config.predict_file, self.utility_config.predict_file_path)
 
-        if key == 'predict':
-            export_data_csv(data, self.utility_config.predict_file, self.utility_config.predict_file_path)
+            logging.info("complete: data ingestion")
+        except ChurnException as e:
+            raise e
